Move Mooc app config to Mooc context
