+++ src/Contexts/Backoffice/Courses/infrastructure/config/production.json
