+++ src/Contexts/Mooc/Shared/infrastructure/config/staging.json
@@ -1,0 +1,1 @@
+{}
