+++ src/apps/mooc/backend/dependency-injection/CoursesCounter/application.yaml
@@ -1,27 +1,27 @@
 services:
   Mooc.coursesCounter.CoursesCounterRepository:
-    class: ../../../../../../Contexts/Mooc/CoursesCounter/infrastructure/persistence/mongo/MongoCoursesCounterRepository
+    class: ../../../../../Contexts/Mooc/CoursesCounter/infrastructure/persistence/mongo/MongoCoursesCounterRepository
     arguments: ["@Shared.ConnectionManager"]
 
   Mooc.coursesCounter.CoursesCounterIncrementer:
-    class: ../../../../../../Contexts/Mooc/CoursesCounter/application/Increment/CoursesCounterIncrementer
+    class: ../../../../../Contexts/Mooc/CoursesCounter/application/Increment/CoursesCounterIncrementer
     arguments: [
       "@Mooc.coursesCounter.CoursesCounterRepository",
       "@Shared.EventBus"
       ]
 
   Mooc.coursesCounter.IncrementCoursesCounterOnCourseCreated:
-    class: ../../../../../../Contexts/Mooc/CoursesCounter/application/Increment/IncrementCoursesCounterOnCourseCreated
+    class: ../../../../../Contexts/Mooc/CoursesCounter/application/Increment/IncrementCoursesCounterOnCourseCreated
     arguments: ["@Mooc.coursesCounter.CoursesCounterIncrementer"]
     tags:
       - { name: 'domainEventSubscriber' }
 
   Mooc.coursesCounter.CoursesCounterFinder:
-    class: ../../../../../../Contexts/Mooc/CoursesCounter/application/Find/CoursesCounterFinder
+    class: ../../../../../Contexts/Mooc/CoursesCounter/application/Find/CoursesCounterFinder
     arguments: ["@Mooc.coursesCounter.CoursesCounterRepository"]
 
   Mooc.coursesCounter.FindCoursesCounterQueryHandler:
-    class: ../../../../../../Contexts/Mooc/CoursesCounter/application/Find/FindCoursesCounterQueryHandler
+    class: ../../../../../Contexts/Mooc/CoursesCounter/application/Find/FindCoursesCounterQueryHandler
     arguments: ["@Mooc.coursesCounter.CoursesCounterFinder"]
     tags:
       - { name: 'queryHandler' }
