+++ src/apps/mooc/backend/dependency-injection/application.yaml
