+++ src/apps/mooc/backend/dependency-injection/application_production.yaml
