+++ src/apps/mooc/backend/dependency-injection/application_staging.yaml
