+++ src/apps/mooc/backend/dependency-injection/apps/application.yaml
@@ -1,13 +1,13 @@
 services:
 
   Apps.mooc.controllers.CoursePutController:
-    class: ../../../controllers/CoursePutController
+    class: ../../controllers/CoursePutController
     arguments: ["@Shared.CommandBus"]
 
   Apps.mooc.controllers.StatusGetController:
-    class: ../../../controllers/StatusGetController
+    class: ../../controllers/StatusGetController
     arguments: []
 
   Apps.mooc.controllers.CoursesCounterGetController:
-    class: ../../../controllers/CoursesCounterGetController
+    class: ../../controllers/CoursesCounterGetController
     arguments: ["@Shared.QueryBus"]
