+++ src/apps/mooc/backend/dependency-injection/index.ts
